Comentarios y Carpetas Fuera
Se documento el código en formato C# y se eliminaron carpetas, el reporte de la investigación referente al FBX Importer se subirá al Trello
