--- Visualizer/Assets/Jhony.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d83571e496e6145459fb9ed1d3834849
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
