--- Visualizer/Assets/Jhony/Materials_Jhony.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cb3cc9f9e4faad247aa56025b36ff80d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
