--- Visualizer/Assets/Jhony/Materials_Jhony/Magenta Light.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 47d22eea7af068d4aa2ad14a21fb684d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
