--- Visualizer/Assets/Jhony/Materials_Jhony/Magenta.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 11b2555632f2d604a8a502a16cbc299e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
