--- Visualizer/Assets/Jhony/Materials_Jhony/Red.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 066e313810ede004b9da8e301d8971d5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
