--- Visualizer/Assets/Jhony/Models_Jhony.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 92c988d28cca0724d960758ea1cb1574
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
