--- Visualizer/Assets/Jhony/Scenes_Jhony.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 15defcc788d5dc54bbca4bc8564c2df6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
