--- Visualizer/Assets/Jhony/Scripts_Jhony.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9d8bb34fb4611ed49a697cf93a9c7d15
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
