--- Visualizer/Assets/Jhony/Scripts_Jhony/DeactivateOnLoad.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5c5666c7e188cc34bb66bbc2527fc53b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
