+++ Visualizer/Assets/Materiales/Checker.meta
