+++ Visualizer/Assets/Materiales/Checker/CheckerMaterial.mat
