+++ Visualizer/Assets/Materiales/Checker/CheckerMaterial.mat.meta
