+++ Visualizer/Assets/Materiales/Checker/tiles_0059_height_4k.png.meta
