+++ Visualizer/Assets/Materiales/Checker/tiles_0059_normal_direct_4k.png.meta
