+++ Visualizer/Assets/Materiales/Checker/tiles_0059_normal_opengl_4k.png.meta
