+++ Visualizer/Assets/Materiales/Checker/tiles_0059_roughness_4k.jpg.meta
